requirements.txt
flask
google-generativeai
PyGithub
python-dotenv